requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
attrs==20.3.0
    # via pytest
certifi==2020.12.5
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.14.4
    # via
    #   -r requirements/base.txt
    #   cryptography
chardet==4.0.0
    # via
    #   -r requirements/base.txt
    #   requests
click==7.1.2
    # via code-annotations
code-annotations==1.0.2
    # via -r requirements/test.in
coverage==5.4
    # via pytest-cov
cryptography==3.2.1
    # via
    #   -r requirements/base.txt
    #   pyjwt
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-filter==2.4.0
    # via -r requirements/base.txt
django-model-utils==4.1.1
    # via
    #   -r requirements/base.txt
    #   edx-organizations
django-simple-history==2.12.0
    # via
    #   -r requirements/base.txt
    #   edx-organizations
django-waffle==2.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   code-annotations
    #   django-crum
    #   django-filter
    #   django-model-utils
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-organizations
    #   rest-condition
djangorestframework==3.12.2
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
    #   edx-organizations
    #   rest-condition
drf-jwt==1.17.3
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-django-utils==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-drf-extensions==6.4.0
    # via
    #   -r requirements/base.txt
    #   edx-organizations
edx-opaque-keys==2.0.2
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   edx-organizations
https://github.com/appsembler/edx-organizations/archive/5.2.0-appsembler11.tar.gz
    # via -r requirements/base.txt
factory-boy==3.1.0
    # via -r requirements/test.in
faker==5.0.0
    # via factory-boy
fs==2.4.12
    # via
    #   -r requirements/base.txt
    #   xblock
future==0.18.2
    # via
    #   -r requirements/base.txt
    #   pyjwkest
idna==2.10
    # via
    #   -r requirements/base.txt
    #   requests
importlib-metadata==2.1.1
    # via
    #   pluggy
    #   pytest
iniconfig==1.1.1
    # via pytest
jinja2==2.11.3
    # via code-annotations
lxml==4.6.2
    # via
    #   -r requirements/base.txt
    #   xblock
markupsafe==1.1.1
    # via
    #   -r requirements/base.txt
    #   jinja2
    #   xblock
mock==3.0.5
    # via -r requirements/test.in
newrelic==5.24.0.153
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==20.9
    # via pytest
pathlib2==2.3.5
    # via pytest
pbr==5.5.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pillow==7.2.0
    # via
    #   -r requirements/base.txt
    #   edx-organizations
pluggy==0.13.1
    # via pytest
psutil==5.8.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
py==1.10.0
    # via pytest
pycparser==2.20
    # via
    #   -r requirements/base.txt
    #   cffi
pycryptodomex==3.9.9
    # via
    #   -r requirements/base.txt
    #   pyjwkest
pyjwkest==1.4.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
pyjwt[crypto]==1.7.1
    # via
    #   -r requirements/base.txt
    #   drf-jwt
pymongo==3.11.2
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r requirements/test.in
pytest-django==4.1.0
    # via -r requirements/test.in
pytest==6.1.2
    # via
    #   -r requirements/test.in
    #   pytest-cov
    #   pytest-django
python-dateutil==2.8.1
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   faker
    #   xblock
python-slugify==4.0.1
    # via code-annotations
pytz==2020.5
    # via
    #   -r requirements/base.txt
    #   django
    #   fs
    #   xblock
pyyaml==5.3.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   xblock
requests==2.25.1
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   pyjwkest
rest-condition==1.0.3
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
semantic-version==2.8.5
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.15.0
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   django-simple-history
    #   edx-drf-extensions
    #   edx-opaque-keys
    #   fs
    #   mock
    #   pathlib2
    #   pyjwkest
    #   python-dateutil
    #   stevedore
    #   xblock
sqlparse==0.4.1
    # via
    #   -r requirements/base.txt
    #   django
stevedore==1.32.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via
    #   faker
    #   python-slugify
toml==0.10.2
    # via pytest
typing==3.7.4.3
    # via
    #   -r requirements/base.txt
    #   fs
urllib3==1.26.3
    # via
    #   -r requirements/base.txt
    #   requests
web-fragments==1.0.0
    # via
    #   -r requirements/base.txt
    #   xblock
webob==1.8.6
    # via
    #   -r requirements/base.txt
    #   xblock
xblock==1.4.0
    # via -r requirements/base.txt
zipp==1.2.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
